% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHP_interactive.R
\name{AHP_interactive}
\alias{AHP_interactive}
\title{Analytic Hierarchy Process}
\usage{
AHP_interactive(criteria, alternatives)
}
\arguments{
\item{{criteria}}{{ a vector of character elements naming the criteria under consideration}}

\item{{alternatives}}{{ a vector of character elements naming the alternatives to be evaluated under the specified criteria}}
}
\value{
A list containing:
\itemize{
\item{a matrix} {which corresponds to the subjective  pairwise criteria comparisons as provided by the user}
\item{a list} {which contains one matrix for each criterion; the matrix provides the subjective pairwise comparisons of the alternatives for the specific criterion as specified by the user }
\item{a vector} {which for each alternative provides the AHP scores computed by the process}
}
}
\description{
Provides for interactive input of subjective user comparisons and suppsequently applies
the Analytic Hierarchy Process (AHP) method of T. L. Saaty in order to obtain ranking
scores for the alternatives.
}
\details{
The function requires the package \code{tcltk} to pop up interactive messages and the
package \code{MCDA} which provides implementation of the \code{AHP} function.
}
\examples{
\dontrun{

## -- Example -- ##
crit <- c("housing", "salary", "career", "recreation")
alt <-c("Stockholm", "Lulea", "Boden", "Malmoe", "Gothenburg")
AHP_interactive(criteria=crit, alternatives=alt)
}


}
\references{
Saaty, T. L. (1980)  The Analytic Hierarchy Process: Planning, Priority Setting, Resource Allocation; McGraw-Hill
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
