% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kraljic_interactive.R
\name{Kraljic_interactive}
\alias{Kraljic_interactive}
\title{Kraljic Matrix}
\usage{
Kraljic_interactive(Items, attribute1, attribute2)
}
\arguments{
\item{{Items}}{{an item catalogue in the form of a vector of character elements (strings)}}

\item{{attribute1}}{{a string holding the name of the first attribute}}

\item{{attribute2}}{{a string holding the name of the second attribute}}
}
\value{
A list containing;
\itemize{
\item\code{Scales} {a data frame containing the limits of the scales for the user estimation of the attributes, as specified by the user}
\item\code{Table} {a data frame containing the user input for each item, the utility scores and the categorization of each of them}
\item\code{Matrix} {a \code{ggplot2} figure of the Kraljic matrix which can be displayed with the \code{print()} function}
}
}
\description{
In purchasing and supply chain management, the Kraljic matrix (or Kraljic model or portfolio) is a method used to
segment the purchases or suppliers of a company by dividing them into four categories; Non-critical,
Bottleneck, Leverage and Bottlemneck based on attributes, typically risk and profitability.
The function provides for interactive input of subjective user estimations of these attributes and subsequently
categorizes purchases or suppliers based on an estimation of a user utility function.
}
\details{
The function requires the package \code{tcltk} to pop up interactive messages and the
package \code{KraljicMatrix} which provides implementation of necessary functions for the
classification of the items and the graphical presentation of the Kraljic portfolio. The latter
requires the package \code{ggplot2} too.
}
\examples{
\dontrun{

## Product names

Items <- c("aaa", "bbb", "ccc", "ddd")

## Attributes

attribute1 <- "Risk"
attribute2 <- "Profitability"

## Run

res <- Kraljic_interactive(Items,attribute1,attribute2)

## Check the scales used

res$Scales

     Risk Profitability
Min    1      1
Max    5     10

## check Table (depending on user input, it can be similar to)

res$Table

   Items Risk Profitability Attr1Score Attr2Score     Category
1   aaa    1     10         0.00000    1.0000000    Strategic
2   bbb    5      1         1.00000    0.0000000 Non-critical
3   ccc    3      5         0.50005    0.6208212     Leverage
4   ddd    3      7         0.50005    0.8096883     Leverage

## Plot the Kraljic matrix

print(res$Matrix)

}




}
\references{
\itemize{
\item Kraljic, P. (1983) Purchasing Must Become Supply Management, Harvard Business Review, September 1983
\item van Weele, A. J. (2014) Purchasing & Supply Chain Management, Cengage.
}
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
