% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDEA_noninteractive.R
\name{LDEA_noninteractive}
\alias{LDEA_noninteractive}
\title{Likert Data Envelopment Analysis (LDEA)}
\usage{
LDEA_noninteractive(Input, Output, Likert = 5)
}
\arguments{
\item{{Input}}{{ a matrix whose rows correspond to DMUs to be evaluated and the columns
to the inputs with respect to which the evaluation is to be performed}}

\item{{Output}}{{ a matrix whose rows correspond to DMUs, the exact same as those of the input and in the same order,
to be evaluated and the columns
to the outputs with respect to which the evaluation is to be performed}}

\item{{Likert}}{{ maximum value of Likert scale in use. Default value is 5.}}
}
\value{
A \code{Table} {which is a data frame completely describing the result; the rows correspond to the evaluated
DMUs and the columns to \code{Theta}  and the DMUs. For an inefficient DMU, i.e., with \code{Theta} < 1,
non-zero values (\code{Lamb}das) on its row indicate the benchmark DMUs. A final column \code{Status}
indicates with a \code{0}-value
whether the corresponding optimization problem has been solved.}
}
\description{
The function implements a Data Envelopment Analysis model/procedure suitable
for data in Likert scale. Likert Data Envelopment Analysis (LDEA) is
based on Mixed Integer Programming (MIP) methodology and is used to measure
the productive efficiency of decision making units (DMUs).
}
\details{
The function requires the package \code{Rglpk} which is based on \code{glpk} and provides
implementation of the necessary functionality for the application of LDEA. The GNU linear programming kit (\code{glpk})
is a Linear and Mixed Integer Programming solver (and much more) and is available to R through the package \code{Rglpk}.
}
\references{
\itemize{
\item Charnes, A., Cooper, W. W. & Rhodes, E. (1978) Measuring the Efficiency of Decision Making Units, European Journal of Operational Research 2 (6) 429–444
\item Charnes, A., Cooper, W. W., Lewin, A. Y.  & Seiford, L. M. (1994) Data Envelopment Analysis: Theory, Methodology, and Applications, Springer
\item Cooper, W. W., Seiford, L. M. &  Tone, K. (2007) Data Envelopment Analysis - A Comprehensive Text with Models, Applications, References and DEA-Solver Software, 2nd Edition, Springer
\item Chen, Y., Cook, W. D., Du, J., Hu, H-. & Zhu, J. (2017) Bounded and Discrete Data and Likert Scales in Data Envelopment Analysis: Application to Regional Energy Efficiency in China, Annals of Operations Research 255, 247-366
}
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
