% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WS_interactive.R
\name{WS_interactive}
\alias{WS_interactive}
\title{The simple Weighted Sum method}
\usage{
WS_interactive(criteria, alternatives)
}
\arguments{
\item{{criteria}}{{ a vector of character elements naming the criteria under consideration}}

\item{{alternatives}}{{ a vector of character elements naming the alternatives to be evaluated under the specified criteria}}
}
\value{
A list containing:
\itemize{
\item\code{Weights} {a vector which contains subjective preference weights of the criteria as specified by the user}
\item\code{Table} {a matrix which corresponds to the subjective performance estimation of the alternatives with respect to the criteria as provided by the user and, in addition, contains the criteria weights as provided by the user, and the scores computed by the weighted sum heuristic for all alternatives}
\item\code{Scores} {a vector which for each alternative provides the weighted sum scores for the alternatives as computed by the process}
\item\code{Order} {a vector which contains the numbering (index) of the alternatives in decreasing order}
}
}
\description{
Provides for interactive input of subjective user judgements and suppsequently applies
the simple Weighted Sum method in order to obtain ranking scores for the alternatives.
}
\details{
The function requires the package \code{tcltk} to pop up interactive messages and the
package \code{MCDA} which provides implementation of the \code{weightedSum} function.
}
\examples{
\dontrun{

## -- Example -- ##

crit <- c("housing", "salary", "career", "recreation")
alt <-c("Stockholm", "Lulea", "Boden", "Malmoe", "Gothenburg")
WS_interactive(criteria=crit, alternatives=alt)

}



}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
